Added a few things, you know how it goes.

+++ PopRPG/Domain/Player.cs
@@ -10,7 +10,7 @@ namespace Domain
     {
         public int Level { get; set; }
         public int Exp { get; set; }
-        public int Hp { get; set; }
+        public double Hp { get; set; }
         public string Armor { get; set; }
         public string Weapon { get; set; }
     }
